+++ core-1.2/src/main/kotlin/kotlin/CharCode1.2.kt
@@ -1,0 +1,5 @@
+// Implement the new post-1.2 APIs which are used by core and serialization
+@file:Suppress("unused")
+package kotlin
+
+inline val Char.code: Int get() = this.toInt()
