+++ core/src/main/kotlin/net/corda/core/flows/CollectSignaturesFlow.kt
@@ -291,8 +291,8 @@ abstract class SignTransactionFlow @JvmOverloads constructor(val otherSideSessio
         // Perform some custom verification over the transaction.
         try {
             checkTransaction(stx)
-        } catch (e: Throwable) {
-            if (e is IllegalStateException || e is IllegalArgumentException || e is AssertionError)
+        } catch (e: Exception) {
+            if (e is IllegalStateException || e is IllegalArgumentException)
                 throw FlowException(e)
             else
                 throw e
