+++ core/src/main/kotlin/net/corda/core/internal/InternalUtils.kt
@@ -52,6 +52,7 @@ import java.security.cert.X509Certificate
 import java.time.Duration
 import java.time.temporal.Temporal
 import java.util.Collections
+import java.util.Locale
 import java.util.PrimitiveIterator
 import java.util.Spliterator
 import java.util.Spliterator.DISTINCT
@@ -622,3 +623,14 @@ val Logger.level: Level
 
 const val JAVA_1_2_CLASS_FILE_FORMAT_MAJOR_VERSION = 46
 const val JAVA_17_CLASS_FILE_FORMAT_MAJOR_VERSION = 61
+
+/**
+ * String extension functions - to keep calling code readable following upgrade to Kotlin 1.9
+ */
+fun String.capitalize() : String {
+    return this.replaceFirstChar { it.titlecase(Locale.getDefault()) }
+}
+fun String.decapitalize() : String {
+    return this.replaceFirstChar { it.lowercase(Locale.getDefault()) }
+}
+
